+++ CHANGELOG.md
@@ -11,7 +11,11 @@ and this project adheres to
 ### Added
 
 - 🔒️(helm) Add pod and container securityContext #1197
-- ✨(summary) add routes v2 for async STT and summary tasks  #1171
+- ✨(summary) add routes v2 for async STT and summary tasks #1171
+
+### Fixed
+
+- 🐛(summary) fix failure webhook notification #1233
 
 ## [1.13.0] - 2026-03-31